refactor: separate modules doc layout, config
