+++ crates/rnote-engine/src/meson.build
@@ -3,7 +3,9 @@ rnote_engine_sources = files(
     'audioplayer.rs',
     'camera.rs',
     'document/background.rs',
+    'document/config.rs',
     'document/format.rs',
+    'document/layout.rs',
     'document/mod.rs',
     'drawable.rs',
     'engine/animation.rs',
